Fix warning
Former-commit-id: 1f7ef270fc5965a65222d9a8e13a0b84448569eb

+++ plugin/fx/include/minko/component/Hologram.hpp
@@ -62,7 +62,7 @@ namespace minko
 			static std::shared_ptr<render::Texture> _backFaceNormalRenderTarget;
 			static std::shared_ptr<render::Texture>	_depthMapRenderTarget;
 
-			static const float MAP_RESOLUTION;
+			static const uint MAP_RESOLUTION;
 
 		public:
